Remove error sampler
