+++ qermit/noise_model/noise_model.py
@@ -10,9 +10,10 @@ from matplotlib.pyplot import subplots  # type: ignore
 # from .pauli_error_transpile import PauliErrorTranspile
 from typing import Dict, Tuple, List
 from pytket.pauli import Pauli
-from .error_sampler import ErrorSampler
+# from .error_sampler import ErrorSampler
 from pytket import Qubit
 from pytket.pauli import QubitPauliString
+from collections import Counter
 
 
 # TODO: For now we should make this explicitly a Pauli error distribution,
@@ -250,7 +251,7 @@ class NoiseModel:
 
         self.noise_model = noise_model
         # self.transpiiler = PauliErrorTranspile(self)
-        self.error_sampler = ErrorSampler(self)
+        # self.error_sampler = ErrorSampler(self)
 
     def reset_seed(self, rng):
         for distribution in self.noise_model.values():
@@ -331,7 +332,7 @@ class NoiseModel:
 
         #     error_counter.update([stabiliser])
 
-        error_counter = self.error_sampler.counter_propagate(
+        error_counter = self.counter_propagate(
             cliff_circ=cliff_circ,
             n_counts=n_rand,
             direction='backward',
@@ -354,3 +355,83 @@ class NoiseModel:
         # print("is_identity_count", is_identity_count)
 
         return error_distribution
+
+    def counter_propagate(self, cliff_circ, n_counts=1000, **kwargs):
+
+        error_counter = Counter()
+
+        # There is some time wasted here, if for example there is no error in
+        # back_propagate_random_error. There may be a saving to be made here
+        # if there errors are sampled before the back propagation occurs?
+        for _ in range(n_counts):
+            stabiliser = self.random_propagate(cliff_circ, **kwargs)
+
+            if not stabiliser.is_identity():
+                error_counter.update([stabiliser])
+
+        return error_counter
+
+    def random_propagate(self, cliff_circ, direction='backward'):
+
+        qubit_list = cliff_circ.qubits
+        stabiliser = Stabiliser(
+            Z_list=[0] * len(qubit_list),
+            X_list=[0] * len(qubit_list),
+            qubit_list=qubit_list,
+        )
+
+        if direction == 'backward':
+            command_list = list(reversed(cliff_circ.get_commands()))
+        elif direction == 'forward':
+            command_list = cliff_circ.get_commands()
+        else:
+            raise Exception(
+                f"Direction must be 'backward' or 'forward'. Is {direction}"
+            )
+
+        for command in command_list:
+
+            if command.op.type in [OpType.Measure, OpType.Barrier]:
+                continue
+
+            if direction == 'forward':
+
+                stabiliser.apply_gate(
+                    op_type=command.op.type,
+                    qubits=command.args,
+                    params=command.op.params,
+                )
+
+            if command.op.type in self.noisy_gates:
+
+                error_distribution = self.get_error_distribution(
+                    optype=command.op.type
+                )
+                error = error_distribution.sample()
+
+                if error is not None:
+                    for pauli, qubit in zip(error, command.args):
+                        if direction == 'backward':
+                            stabiliser.pre_apply_pauli(
+                                pauli=pauli, qubit=qubit)
+                        elif direction == 'forward':
+                            stabiliser.post_apply_pauli(
+                                pauli=pauli, qubit=qubit)
+                        else:
+                            raise Exception(
+                                "Direction must be 'backward' or 'forward'. "
+                                + f"Is {direction}"
+                            )
+
+            if direction == 'backward':
+
+                # Note that here we wish to pull the pauli back through the gate,
+                # which has the same effect on the pauli as pushing through the
+                # dagger.
+                stabiliser.apply_gate(
+                    op_type=command.op.dagger.type,
+                    qubits=command.args,
+                    params=command.op.dagger.params,
+                )
+
+        return stabiliser
